2058: Put tech pages into their own app
