+++ accessibility_monitoring_platform/apps/cases/utils.py
@@ -3,18 +3,13 @@ Utility functions for cases app
 """
 
 import copy
-import csv
-import io
 import logging
-from datetime import datetime, timezone
 from typing import Any
 
-from django.contrib.auth.models import User
 from django.db.models import Case as DjangoCase
 from django.db.models import Q, QuerySet, When
 
 from ..common.utils import build_filters, extract_domain_from_url
-from ..detailed.models import DetailedCase, DetailedCaseHistory
 from ..simplified.models import SimplifiedCase
 from .models import CASE_STATUS_UNASSIGNED, BaseCase, Sort
 
@@ -109,57 +104,3 @@ def find_duplicate_cases(url: str, organisation_name: str = "") -> QuerySet[Base
             Q(organisation_name__icontains=organisation_name) | Q(domain=domain)
         )
     return BaseCase.objects.filter(domain=domain)
-
-
-COMMENT_FULLNAME_TO_USERNAME: dict[str, str] = {
-    "amy.wallis": "amy.wallis@digital.cabinet-office.gov.uk",
-    "Andrew Hick": "andrew.hick@digital.cabinet-office.gov.uk",
-    "ChrisH": "chris.heathcote@digital.cabinet-office.gov.uk",
-    "Eu-Hyung Han": "eu-hyung.han@digital.cabinet-office.gov.uk",
-    "Jessica Eley": "jessica.eley@digital.cabinet-office.gov.uk",
-    "katherine.badger": "katherine.badger@digital.cabinet-office.gov.uk",
-    "Keeley Robertson": "keeley.talbot@digital.cabinet-office.gov.uk",
-    "Kelly Clarkson": "kelly.clarkson@digital.cabinet-office.gov.uk",
-    "nesha.russo": "nesha.russo@digital.cabinet-office.gov.uk",
-    "Nicole Tinti": "nicole.tinti@digital.cabinet-office.gov.uk",
-}
-TRELLO_COMMENT_LABEL: str = "Imported from Trello"
-
-
-def import_trello_comments(csv_data: str) -> None:
-    try:
-        comment_fullname_to_user: dict[str, User] = {
-            comment_fullname: User.objects.get(username=username)
-            for comment_fullname, username in COMMENT_FULLNAME_TO_USERNAME.items()
-        }
-    except User.DoesNotExist:  # Automated tests
-        return
-    DetailedCaseHistory.objects.filter(label=TRELLO_COMMENT_LABEL).delete()
-    reader: Any = csv.DictReader(io.StringIO(csv_data))
-    for row in reader:
-        case_no: str = row["case_no"]
-        if case_no == "":
-            continue
-        if row["case_no"].startswith("D"):
-            case_identifier: str = f"#D-{case_no[1:]}"
-            try:
-                detailed_case: DetailedCase = DetailedCase.objects.get(
-                    case_identifier=case_identifier
-                )
-                detailed_case_history: DetailedCaseHistory = (
-                    DetailedCaseHistory.objects.create(
-                        detailed_case=detailed_case,
-                        event_type=DetailedCaseHistory.EventType.NOTE,
-                        value=row["comment_text"],
-                        label=TRELLO_COMMENT_LABEL,
-                        created_by=comment_fullname_to_user[row["comment_fullname"]],
-                    )
-                )
-                comment_time: datetime = datetime.strptime(
-                    row["comment_date"], "%Y-%m-%dT%H:%M:%S.%fZ"
-                ).replace(tzinfo=timezone.utc)
-                detailed_case_history.created = comment_time
-                detailed_case_history.updated = comment_time
-                detailed_case_history.save()
-            except DetailedCase.DoesNotExist:
-                logger.warning("DetailedCase not found: %s", case_identifier)
