+++ accessibility_monitoring_platform/apps/tech/admin.py
@@ -1,0 +1,3 @@
+"""
+No admin for dashboard app
+"""
