+++ accessibility_monitoring_platform/apps/tech/models.py
@@ -1,0 +1,3 @@
+"""
+No models for dashboard app
+"""
