+++ accessibility_monitoring_platform/apps/tech/templates/tech/import_trello_comments.html
@@ -33,7 +33,7 @@
                                 class="govuk-button"
                                 data-module="govuk-button"
                             />
-                            <a href="{% url 'common:reference-implementation' %}" class="govuk-link govuk-link--no-visited-state">
+                            <a href="{% url 'tech:reference-implementation' %}" class="govuk-link govuk-link--no-visited-state">
                                 Reference implementation</a>
                         </div>
                     </div>
