+++ accessibility_monitoring_platform/apps/tech/templates/tech/ref/widgets.html
@@ -1,13 +1,13 @@
 <div class="govuk-grid-row">
     <div class="govuk-grid-column-full">
         <h2 id="id_widgets" class="govuk-heading-m">Widgets</h2>
-        {% include 'common/tech_team/ref_widget_markdown.html' %}
+        {% include 'tech/ref/widget_markdown.html' %}
         <hr class="amp-width-100" />
-        {% include 'common/tech_team/ref_widget_go_back.html' %}
+        {% include 'tech/ref/widget_go_back.html' %}
         <hr class="amp-width-100" />
-        {% include 'common/tech_team/ref_widget_search.html' %}
+        {% include 'tech/ref/widget_search.html' %}
         <hr class="amp-width-100" />
-        {% include 'common/tech_team/ref_widget_copy_text.html' %}
+        {% include 'tech/ref/widget_copy_text.html' %}
         <hr class="amp-width-100 amp-margin-bottom-30" />
     </div>
 </div>
