+++ accessibility_monitoring_platform/apps/tech/templates/tech/reference_implementation.html
@@ -20,18 +20,18 @@
                 <p class="govuk-body">Examples and links to components and patterns used in the platform.</p>
             </div>
         </div>
-        {% include 'common/tech_team/ref_contents.html' %}
-        {% include 'common/tech_team/ref_banners.html' %}
-        {% include 'common/tech_team/ref_forms.html' %}
-        {% include 'common/tech_team/ref_formlike.html' %}
-        {% include 'common/tech_team/ref_gds_components.html' %}
-        {% include 'common/tech_team/ref_headers.html' %}
-        {% include 'common/tech_team/ref_labels.html' %}
-        {% include 'common/tech_team/ref_links.html' %}
-        {% include 'common/tech_team/ref_lists.html' %}
-        {% include 'common/tech_team/ref_page_layouts.html' %}
-        {% include 'common/tech_team/ref_tables.html' %}
-        {% include 'common/tech_team/ref_widgets.html' %}
+        {% include 'tech/ref/contents.html' %}
+        {% include 'tech/ref/banners.html' %}
+        {% include 'tech/ref/forms.html' %}
+        {% include 'tech/ref/formlike.html' %}
+        {% include 'tech/ref/gds_components.html' %}
+        {% include 'tech/ref/headers.html' %}
+        {% include 'tech/ref/labels.html' %}
+        {% include 'tech/ref/links.html' %}
+        {% include 'tech/ref/lists.html' %}
+        {% include 'tech/ref/page_layouts.html' %}
+        {% include 'tech/ref/tables.html' %}
+        {% include 'tech/ref/widgets.html' %}
     </main>
 </div>
 {% endblock %}
