+++ stack_tests/integration_tests/cypress/e2e/axe.cy.js
@@ -75,13 +75,13 @@ const loggedInUrls = [
   '/common/metrics-report/',
   '/common/platform-versions/',
   '/common/report-issue/',
-  '/common/issue-reports-list/',
-  '/common/platform-checking/',
-  '/common/reference-implementation/',
   '/notifications/cases/1/reminder-task-create/',
   '/notifications/task-list/',
   '/reports/edit-report-wrapper/',
   '/reports/1/report-metrics-view/',
+  '/tech/',
+  '/tech/issue-reports-list/',
+  '/tech/platform-checking/',
   '/user/1/edit-user/'
 ]
